+++ src/main/kotlin/test/ktortemplate/core/persistance/sql/CarRepositoryImpl.kt
@@ -9,10 +9,10 @@ import org.koin.core.inject
 import test.ktortemplate.conf.database.DatabaseConnection
 import test.ktortemplate.core.model.Car
 import test.ktortemplate.core.model.CarSaveCommand
-import test.ktortemplate.core.model.PageRequest
 import test.ktortemplate.core.persistance.CarRepository
 import test.ktortemplate.core.utils.createFromFilters
 import test.ktortemplate.core.utils.createSorts
+import test.ktortemplate.core.utils.pagination.PageRequest
 
 class CarRepositoryImpl : CarRepository, KoinComponent {
