+++ src/main/kotlin/test/ktortemplate/core/service/CarService.kt
@@ -2,7 +2,7 @@ package test.ktortemplate.core.service
 
 import test.ktortemplate.core.model.Car
 import test.ktortemplate.core.model.CarSaveCommand
-import test.ktortemplate.core.model.PageRequest
+import test.ktortemplate.core.utils.pagination.PageRequest
 
 interface CarService {
     fun count(pageRequest: PageRequest): Int
