+++ src/main/kotlin/test/ktortemplate/core/utils/pagination/PageRequest.kt
@@ -1,4 +1,4 @@
-package test.ktortemplate.core.model
+package test.ktortemplate.core.utils.pagination
 
 enum class SortFieldOrder {
     asc,
